Fix Action Mailer not skipped when Active Job is
Previously, generating an app with `rails new jobless --skip-active-job`
would error because it would still generate files and configuration for
Action Mailer. The Action Mailer railtie would be commented in
config/application because it correctly used the skip_action_mailer?
method added in 9578865, so the app generation would fail when it tried
to reload the app.

This commit fixes the issue by using skip_action_mailer? everywhere
instead of options[:skip_action_mailer], which isn't aware of the Active
Job dependency.

This fix is only necessary for 7-0-stable because --skip-* option
dependencies were completely reworked on main in 4c67975. This rework
removed the need for the skip_*? methods and made the values in
options[:skip.*] always correct.