Fix double logs if Rails.logger outputs to STDERR
Previously, configuring Rails.logger to output to STDERR would result in
double logs when running the `rails server` command. This was due to the
server command broadcasting the logger to STDOUT if it did not already
output to STDOUT.

This commit fixes the issue by checking that the logger does not output
to STDERR in addition to STDOUT before broadcasting. This aligns it with
a similar broadcast done in the Active Record railtie.