Fix test where broadcaster doesn't define #silence
This test was added in 308e84e to ensure that calling #silence on a
broadcasted logger works if the broadcastee implements #silence but the
broadcaster does not.

However, `include ActiveSupport::LoggerSilence` was moved from the
`FakeLogger` to the `CustomLogger` in 05ad44e. This resulted in both
loggers in the above test implementing silence, so it no longer tested
the behavior expected (that only one of them implements #silence).

The inclusion of `LoggerSilence` was moved back to `FakeLogger` so that
`CustomLogger` does not implement silence, and assertions were added so
that this does not regress. In addition, methods transitively included
by `LoggerSilence` are overriden so that they continue to behave like
`CustomLogger`.