Clarify missing template error page
Some goals here:
- Be less obtuse with word choices for folks who use English
as a second language.
- Stop referring to view templates as "templates" when the
literal directory is called `app/views` which has and will
confuse beginners (otherwise let's rename the directory).
- Stop formatting the HTML with `<br>` tags like it's 2005
and let the text flow with the viewport as appropriate.
- Give a *concrete* example of the naming relationship between
a controller and its view template, based on
[the ActionView Rails Guides][1].

I also tried to clarify the frankly confusing copy explaining
the 204 No Content rendering case, which I think should
probably mention that such endpoints shouldn't be allowed to
render HTML to avoid this error, but that might be pushing
the scope of the error feedback a bit too far.

[1]: https://guides.rubyonrails.org/layouts_and_rendering.html