Don't remove duplicates from args array
The [pipe operator](https://ruby-doc.org/core-3.1.2/Array.html#method-i-7C) will return the union of the two sets, with duplicates removed. That causes problems if there are two values in the arguments that are the same.

This change still avoids adding a duplicate `--skip-bootsnap` option, but doesn't remove duplicates which allows for args like ["--option1", "value", "--option2", "value"] to work as expected.

Fixes #46347