Fix invalid content type deprecation warning
This fixes the following warning when running Active Storage tests:

  ```
  DEPRECATION WARNING: image/jpg is not a valid content type, it should not be used when creating a blob, and support for it will be removed in Rails 7.1. If you want to keep supporting this content type past Rails 7.1, add it to `config.active_storage.variable_content_types`. Dismiss this warning by setting `config.active_storage.silence_invalid_content_types_warning = true`.
  ```

Note that this test should eventually be removed.  It was written to
test the fix from #42225; however, after the deprecation from #42227 is
complete and invalid content types have been removed from
`config.active_storage.variable_content_types`, calling `variant` when
the content type is invalid will raise `ActiveStorage::InvariableError`
(instead of the `NoMethodError` mentioned in #42225 / #41777).  And that
behavior is already tested by the "variation of invariable blob" test.