Add missing defaults to new_framework_defaults
These were added in cdce275, 5d0c2b0, and 5256c90

For the new MessageVerifier and MessageEncryptor defaults, I modeled the
documentation after cb76ede. The goal is to warn users that they should
pay attention to the config and not just uncomment it without doing
additional work.

I also added an extra note to log_file_size since it is the first
load_default I'm aware of that is environment dependent.

Co-authored-by: Jonathan Hefner <jonathan@hefner.pro>