Only ask people to run migration in a environment in non-development environments
When running test, you might have pending migrations. Before Rails would
tell you to run migrations with RAILS_ENV=test.

This causes the problem that now the development environment still have
pending migrations.

The correct workflow is to run migrations in the development environment
and let Rails prepare the test database based on the schema or structure
file.

In the development environment we don't need to tell to use the
`RAILS_ENV` environment variable.