Map through subtype in PostgreSQL::OID::Array
This changes `PostgreSQL::OID::Array#map` to invoke `subtype.map` for
each element of the array.  When `subtype` is a scalar, `subtype.map`
(as implemented by `ActiveModel::Type::Value#map`) will simply call the
block with the element.  When `subtype` is a non-scalar, such as a
`PostgreSQL::OID::Range`, `subtype.map` ensures that the nested scalar
values are properly mapped.

Partially reverts #45348, as an alternative fix.
Fixes #46371.
Closes #46372.

Co-authored-by: fatkodima <fatkodima123@gmail.com>