Fix collation for changing column to non-string
3a0d0f4 made change_collumn preserve the existing column's collation.
However, this leads to problems when the old and new column types have
incompatible collations.

be0a58b addressed this by adding a check to ensure that the new column
type isn't binary. However, this did not cover every column type
impacted by this issue.

008d9d1 followed up by changing the collation to only be preserved if
both the old and new column types are string-like, fixing the issue for
other column types like :int or :blob.

This is a manual backport of 008d9d1 because it also touched some other
code only present on main.

Co-authored-by: Andrew Hoglund <ahoglund@github.com>