Add link to Rollback in transactions doc.
A common misconception with transaction rollback in Rails is that
exceptions will not be re-raised from nested transactions. This is only
true for ActiveRecord::Rollback. The documentation for that class makes
that clear.

A link to that class doc has been added, as well as an explicit
restatement that all other exceptions will be re-raised.