Fix an issue where changes will disappear.
When `autosave: true` is set on both ends of a relation `save` and/or `valid?` can be called more than once removing the changes that just happened during the current `save` and making `previous_changes` and empty hash.

This commit adds an instance variable on Connection to ensure that each record only gets `save` and/or `valid?` called once while traversing the autosave tree, ensuring the changes are available after saving via the `previous_changes` method.

Fixes #28491