Revert "Merge pull request #46282 from jonathanhefner/active_model-forgetting_assignment-avoid-value_for_database"
This reverts commit 1f039d8f400e2b9bf76a9c25bd3916eb0aefa0f7, reversing
changes made to be0b5c65a175b6c92514375fc7044efb11cdbe90.

This revert is temporary while we debug some issues with our app. While
we're pretty sure this is caused by code in our application and a
compbination with `activerecord-typed_store`, the failure mode is easy
to miss because it doesn't raise an exception.

We will un-revert after a bit more investigation. We just want to be
confident that the other cases where this could cause issues are fixed
without being blocked from upgrading for the next month or so.