Add filtering of encrypted attributes in #inspect
Previously, encrypted attributes could be added to an application's
filter_parameters which would filter the attribute values from logs.

This commit makes the add_to_filter_parameters additionally add
encrypted attributes to records' filter_attributes, which allows them
to be filtered when models are inspected (such as in the console).