Update deprecation-related config descriptions [ci-skip]
This updates the description of several deprecation-related
configuration settings to clarify that their "default" values are from
the default generated `config/environments` files, and that those values
vary by environment.  This also updates the description of the
`active_support.deprecation_behavior` initializer to mention the new
`Rails.application.deprecators`, and to link to the relevant
configuration settings.