Add support for reversing FIELD order
No longer raise ActiveRecord::IrreversibleOrderError when calling
`reverse_order` on an ActiveRecord::Relation that has a `FIELD`
ordering. Instead, flip the direction that the `FIELD` ordering
is sorted in, to reverse the order of the results that will be
returned by the query.

Using a subexpression call requiring Ruby 2.0, see https://rotadev.com/regular-expression-to-match-balanced-parentheses-dev/.