Fix undefined attribute method with attr_readonly
The problem is where an abstract class defines a readonly attribute.
When another class subclasses that parent with
`raise_on_assign_to_attr_readonly`, it will define the readonly method
before the abstract parent has defined its attributes. As a result, when
those methods are defined, they will not define the readonly attribute
writer, so that `super` called from the readonly accessor will fail with
a `NoMethodError`.

Instead of trying to make the overriden attribute method lazier by
defining it after the real attribute method has been defined, this
commit changes the method override to be one layer down on
_write_attribute. This avoids the issue of the attribute method being
undefined because the real _write_attribute will always be defined.

Co-authored-by: Adrianna Chang <adrianna.chang@shopify.com>
Co-authored-by: Chris Salzberg <chris.salzberg@shopify.com>