Update migration generator timestamp behavior
Currently, the documentation says that the generator appends a timestamp to the file name (line: 137), however, the generator prepends the timestamp to the file name as noted on line 111 where it shows `YYYYMMDDHHMMSS_create_products.rb`. This PR updates the guides to more accurately reflect the behavior of where the generator adds the timestamp to the file name.