Clarify load_environment_config initializer order [ci-skip]
The description of `load_environment_hook` does mention that
`load_environment_config` runs before it, but that description is at the
distant top of the list.  This commit makes it clear that
`load_environment_config` runs before `load_environment_hook` without
having to read the list from top to bottom.