Add failing test
The change in https://github.com/rails/rails/pull/46105 introduces a
subtle failing edge case, which this test reproduces.

The problem is where an abstract class defines a readonly attribute.
When another class subclasses that parent with
`raise_on_assign_to_attr_readonly`, it will define the readonly method
before the abstract parent has defined its attributes. As a result, when
those methods are defined, they will not define the readonly attribute
writer, so that `super` called from the readonly accessor will fail with
a `NoMethodError`.