Memoize the value of show_exceptions?
Previously, the value of `action_dispatch.show_exceptions` would be
compared with `false` every time `show_exceptions?` is called. As
described in the previous comment, the purpose of the comparison is to
treat a `nil` configuration as `true`.

This commit does what the rest of the comment suggests, which is to move
the comparison into `env_config` so that `show_exceptions?` just looks
up the value stored in `env` and doesn't have to do any comparisons.