Clarify CI requirement in PR template [ci skip]
You can't check if "CI is passing" until you create a PR, because for most people Buildkite will only run the build when a PR is created (It will run on branch push only if you push to a branch of the `rails/rails` repo, which only people with write access can do.). So we shouldn't tell people to wait for CI to pass before submitting a PR.

This PR just tweaks the CI template to explain this in a comment so that people know to check back once CI has actually run.