Use NameError for missing association model class
Follow-up to #46605.

`AssociationReflection#automatic_inverse_of` expects a `NameError` to be
raised when a model class is missing.  #46605 changed the error to an
`ArgumentError`, which caused failures in isolated Active Record tests:

* https://buildkite.com/rails/rails/builds/91327#0184c50a-a38c-496e-b877-2946a3966cd9/1453-1463
* https://buildkite.com/rails/rails/builds/91327#0184c50a-a38e-4243-9b7c-82ff1728216a/1438-1448
* https://buildkite.com/rails/rails/builds/91327#0184c50a-a38e-4243-9b7c-82ff1728216a/1596-1606

Additionally, other `MacroReflection` subclasses raise a `NameError`,
not an `ArgumentError`, when the relevant class is missing.  (See
`MacroReflection#compute_class`.)

This commit changes the error back to `NameError`, and adds a regression
test to catch the problem even when the tests are not isolated.

This commit also reverts #46616 and #46619, which were only to address
the failures in the isolated tests.  For additional context, see #32113.