Better `enable_reloading` comment in `environments/test.rb`
In the process of removing spring, https://github.com/rails/rails/pull/42997 added a comment for `config.cache_classes`. Through that PR and subsequent revisions the comment has gotten a bit confusing - it doesn't explain what the `enable_reloading` setting does, just how to revert it if you are using Spring (which isn't included in new apps anymore).

So this PR just updates the template with a comment about what the setting does. If someone really wants to use Spring there's other places to learn about how to change this setting, as that's not a default path we're encouraging anymore.