Fix protected envs multi-db test
Locally, this test was failing because the configurations for the test
environment were empty. If they are empty, we won't raise the
appropriate error.

In the Rails test suite locally the environment in this test is
`default_env`, but in CI it's `test`. We can fix this by pass `env` to
`check_protected_environment!` instead of passing a string of the
enviornment.

Followup to #46340