Fix internal_metadata connection
Previously, this code was using `ActiveRecord::Base.connection` for the
internal metadata. However, if a MigrationContext was initialized with a
different connection's internal metadata we should use that. This fixes
the `record_environment` method to use the correct connection for
storing the environment. We noticed this in flaky tests at Shopify
causing internal metadata to get written to the wrong readonly
connection. I have added a test so we don't miss this in the future.