Run `test:prepare` before `bin/rake test` command
Alternative approach to https://github.com/rails/rails/pull/46664. This PR tries to tackle the same problem, but does a bit less in the implementation.

Here, we only modify `bin/rake test` to run `rake test:prepare`. This allows us to have one command that will reliably run all tests and run asset bundling (or anything else that's a prerequisite for tests) in a CI environment.

I've also updated the Testing with Rails guide with instructions on how to run tests in CI.