Allow SQL warnings to be reported for MySQL and Postgres.
`ActiveRecord.db_warnings_action` can be used to configure the
action to take when a query produces a warning. The warning can be
logged, raised, or trigger custom behaviour provided via a proc.

`ActiveRecord.db_warnings_ignore` allows applications to set an
allowlist of SQL warnings that should always be ignored, regardless
of the configured action.

Co-authored-by: Paarth Madan <paarth.madan@shopify.com>