Fix parallel testing databases
In #46270 it was reported that there were database errors when running
the tests in an app. I set up a demo app to use parallel testing and was
able to reproduce. The issue was that the `reconstruct_from_schema`
method needs to use the `pool` and not the `connection` because the
database might not exist yet and will raise an error. I don't know how
to test this inside Rails but I verified this behavior in my demo app.