Remind of credentials diff enrollment on edit
When users create a new app or `git clone` an existing app, they may not
realize it is enrolled in decrypted diffing of credentials (or be aware
of the feature).

This commit adds a confirmation message when configuring the Git diff
driver for credentials, to serve as a reminder.  The message will be
shown when running `bin/rails credentials:edit` for the first time after
`rails new` or `git clone`.

This commit also silences the enrollment confirmation message when
creating a new app, to reduce noise.