YAML by default now uses safe loading and only accepts a limited number of classes which includes Hash, but not HashWithIndifferentAccess.
When using accepts_nested_attributes_for with a serialized Hash attribute,
a `Psych::DisallowedClass: Tried to load unspecified class: ActiveSupport::HashWithIndifferentAccess`
error would be thrown even if the nested attribute is a regular hash.

This patch forces the nested attribute to a regular Hash attribute when
using a serialized Hash. In general, the nested attributes are converted
to a HashWithIndifferentAccess for processing, but this breaks the safe
loading for YAML which doesn't accept HashWithIndifferentAccess, so the
serialized attribute must be forced to a Hash.