Add CMD and refactor ENTRYPOINT
The CMD will run by default when you run the container without supplying
any arguments but you can choose to overwrite this at runtime by
doing: docker container run -it myapp rails c

The ENTRYPOINT will only create or migrate the database when the default
CMD is invoked. If you choose to run a different command such as the
rails console or a background worker process the db:prepare command
will not run.