Only register CurrentAttributes hooks if used
Previously, executor hooks were always defined to clear
ActiveSupport::CurrentAttributes whether or not an application defines
a Current class. However, those apps without CurrentAttributes will end
up autoloading the class when the executor hooks run, which adds
additional overhead to the first request or job processed.

One possible solution would be to eager load CurrentAttributes to
prevent it from being autoloaded mid request, however this would be
extraneous for applications not using it. Therefore a better solution
is to only register the executor hooks when CurrentAttributes is loaded.