Change dockerfile from using Node 19 to match dev environment
This is accomplished via installing volta: https://volta.sh/.

Also:
  * decouple the installation of libvips from node
  * add installation of the node_modules by running yarn

This pull request is part one of many intended to add features
from fly.io's generated Dockerfiles that are of general interest
to Rails.  For the overall plan, see:
https://community.fly.io/t/preparations-for-rails-7-1/9512

I'm making these changes in stages so that I can get early feedback
and in so doing improve future pull requests.

This changes was tested by generating an new application with esbuild
and deploying it.

attn: @ddh