Eager load Rails::BacktraceCleaner
Previously, Rails::BacktraceCleaner was both lazily required and
instantiated. While it would end up being loaded during initialization
for apps that use Active Record (due to active_record.backtrace_cleaner
initializer), it would not be loaded until the first request for those
that don't.

The solution is similar to daff36c, however ::Rails is not currently
in the list of eager_load_namespaces. Therefore simply requiring it
seems like the easiest solution.

The removed comment indicating the lazy require is necessary due to load
order dates back to the addition of Rails::BacktraceCleaner in f42c77f.
This predates the integration of Bundler and reorganization of framework
loading, so it is no longer accurate.