Make belongs_to validate the presence of the foreign key
By default, `belongs_to` automatically validates the presence of the
association. For example, given the following model:

```
class Article < ApplicationRecord
  belongs_to :author
end
```

Rails will validate Article.author is non-nil. Good!

However, per the Rails guides, one should add validation on the *_id
foreign key field as these are the fields rendered in HTML forms. For
examples, see the guides:

https://guides.rubyonrails.org/form_helpers.html (look for :city_id).

https://guides.rubyonrails.org/active_record_validations.html#custom-methods (look
for :customer_id).

As of right now, if `author_id` is rendered in an HTML form using a
`select` and the user fails to choose an option, the default validation
doesn't render a message to the user. Instead, they receive now error.

To workaround this, applications must manually add
`validate_presence_of` (or similar) to `belongs_to` foreign key that
will be rendered in forms. We can avoid this repetitive code and make
the validation work more smoothly with forms.

To solve the above, when adding validation on the belongs_to
association, add it to the foreign key at the same time.