Add foreign key constraints warning about `destroy_async`
It is not possible to use the `dependent: : destroy_async` option on
associations when they are backed by foreign key constraints. There
is a warning on the Rails API about this, but there is currently no
mentions of this problem on the Rails guides - which is the first page
found on search engines when searching about this option.

This commit copies the same warning found on the Rails API to the
Rails guides, making this restriction more visible.

[ci skip]