Document AD::Request::RFC* constants individually [ci-skip]
Prior to this commit, the list of RFCs and URLs was jumbled in a single
paragraph (with no delimiters) that was associated with the `RFC2616`
constant.

This commit splits up the list, associating each RFC with its respective
constant, and incorporating each URL into a link.

(cherry picked from commit 1aba8f64bf507724715e488a6936866680f7aa0d)