Split Templates, Partials and Layouts sections in guide [ci-skip]
Templates, Partials and Layouts deserve their own sections.
All headings of these section can now be moved a level up, making these
headings show up in the sidebar section. This improves the overview of
what is described in the guide.

Partial Layouts, which already had their own section, is now moved under
the Layouts section, as it is part of the layouts.