Improve Grouping headings in Active Record Querying guide [ci-skip]
As `GROUP BY` is the actual clause and not `GROUP`, we could rename the
`Group` heading to `Group By`. But as other headings (Ordering, Selecting,
Joining) describe the usage instead of the SQL statement, it has been
renamed to it's usage: `Grouping`.

The `HAVING` condition requires a `GROUP BY` condition, so it's moved
under the `Grouping` section. The heading has been renamed similar to
other headings like `NOT conditions.