Fix tests where assertions assume Selenium 4.x
In 3dbef44, the assertions were changed and a minimum version was added
for selenium-webdriver. The version was later changed in 2a6385a because
the assertions required changing again but the new selenium-webdriver
version did not have the same Ruby version support as Rails 6-1-stable.

The issue is that by not specifying a prerelease in the version
constraint, Bundler did not know we wanted a prerelease. Since the last
version of webdrivers to support Ruby 2.5 requires selenium-webdriver <
4.0, this meant that it would choose a 3.x version instead.

By adding back the minimum version constraint specifying a prerelease,
Bundler can now choose a 4.x prerelease version of selenium-webdriver
and the assertions will once again pass.