Manually backport #46933
I wasn't able to cherry-pick this without conflicts so I manually
re-implemented the changes.

Original message:

Fix bundle install broken in Ruby 2.5 image builds

There are two issues being fixed:

The first is that sneakers 2.12.0 requires rake 12, and bundler has
trouble resolving this. This was reported as an issue in bundler
2.4.0-dev but appears to also affect 2.3.26 somehow. Since bundler 2.4+
no longer supports Ruby 2.5 it seems easier to relax the rake version
on Ruby 2.5 than try to fix bundler.

The other is a continuation of the net-http issues. mail 2.8.0 defines
net-smtp, net-imap, and net-pop as dependencies. All of these gems
depend on the net-protocol gem. There's an issue in Ruby <= 2.7 where
the built in net-http uses require_relative to load built in
net-protocol.

To prevent warnings from net-protocol being loaded from two different
places, we add net-http as an extra gem dependency so that the built in
net-http is never required (so that built in net-protocol is never
required).

However, this workaround is impossible on Ruby 2.5 as net-http 0.1.0 has
a minimum ruby version of 2.6.0. So to prevent redefinition warnings, we
can lock mail so that the gem version of net-protocol is never required.

Finally, strscan needs to be added to the Gemfile because it was
previously a dependency of net-imap and the bug report templates need
certain gems (like strscan) installed before they run or they will error
out. See commits like 76ac6e9, 48c9b48, and 2e8381c for related fixes.

Fix tests where assertions assume Selenium 4.x

In 3dbef44, the assertions were changed and a minimum version was added
for selenium-webdriver. The version was later changed in 2a6385a because
the assertions required changing again but the new selenium-webdriver
version did not have the same Ruby version support as Rails 6-1-stable.

The issue is that by not specifying a prerelease in the version
constraint, Bundler did not know we wanted a prerelease. Since the last
version of webdrivers to support Ruby 2.5 requires selenium-webdriver <
4.0, this meant that it would choose a 3.x version instead.

By adding back the minimum version constraint specifying a prerelease,
Bundler can now choose a 4.x prerelease version of selenium-webdriver
and the assertions will once again pass.