Support composite foreign key association assignments
Given a model with a composite primary key, it is possible to assign
the association associated with the model by the composite foreign key.

For example, given two `Sharded::BlogPost` and `Sharded::Comment` models
that are associated using a composite foreign key so that:
`Sharded::BlogPost.has_many :comments, foreign_key: [:blog_id, :blog_post_id]`,
it is possible to assign an instance of `Comment` to an instance of
`BlogPost` like `blog_post.comments << comment`.
The same applies to the opposing `:belongs_to` association.