Mention that Rails will autoload validators
I don't write custom validators often, but have read this doc many times and it never occurred to me that validators could be autoloaded if placed in `app/validators` until I stumbled on it when trying to figure out how to "register" them (e.g. `EmailValidator` to `email:`). (Separately, it might be worth mentioning explicitly that the kwarg is derived from the class name).